inversed lidar scan angles, fixed nasty imu bug, pixhawk mocker,  plot lidar and imu together, wrote start of line merge function

+++ poseEstimationBasedOnLidar/pixhawkWrapper.py
@@ -10,10 +10,12 @@ from multiprocessing import Process, Pipe
 import matplotlib.pyplot as plt
 import numpy as np
 
+READ_FROM_SERIAL = 1
+READ_FROM_FILE = 2
 
 def callbackIMU(data, pipeToMain):
     global yaw
-    yaw += data.angular_velocity.z
+    yaw += data.angular_velocity.z*0.02
     pipeToMain.send(yaw)
 
 #parallel process needed because imu data was not being fetched at 50Hz, because main was blocking thread
@@ -22,7 +24,7 @@ def parallelFuncImuYaw(pipeToMain):
     yaw = 0
     rospy.init_node('listener', anonymous=True)
     #tcp nodelay, don wait for ACK, just send all pakcets individually asap
-    rospy.Subscriber("/mavros/imu/data", Imu, callbackIMU, pipeToMain, tcp_nodelay=True)
+    rospy.Subscriber("/mavros/imu/data_raw", Imu, callbackIMU, pipeToMain, tcp_nodelay=True)
     while 1:
         sleep(0.01)
     pipeToMain.close()
@@ -33,30 +35,38 @@ def parallelFuncImuYaw(pipeToMain):
 #     return (atan2(2.0 * (q[3] * q[0] + q[1] * q[2]) , - 1.0 + 2.0 * (q[0] * q[0] + q[1] * q[1])))
 
 class pixhawk():
-    def __init__(self):
+    def __init__(self, readFrom):
         self.previousYaw = None
         self.orientation = None
         self.start_sec = time()
 
-        self.pipeFromImu, child_conn = Pipe()
-        self.processPipeImuYaw = Process(target=parallelFuncImuYaw, args=(child_conn,))
-        self.processPipeImuYaw.start()
+        self.readFrom = readFrom
+        if readFrom == READ_FROM_SERIAL:
+            self.pipeFromImu, child_conn = Pipe()
+            self.processPipeImuYaw = Process(target=parallelFuncImuYaw, args=(child_conn,))
+            self.processPipeImuYaw.start()            
 
     def closeParallelProcess(self):
-        self.processPipeImuYaw.terminate()
+        if self.readFrom == READ_FROM_SERIAL:
+            self.processPipeImuYaw.terminate()
     
     def getImuYawDisplacement(self):
-        yaw = None        
-        while self.pipeFromImu.poll():
-            yaw = self.pipeFromImu.recv()
-        if yaw is not None:
-            print(yaw)
-            if self.previousYaw is not None:
-                imuYawBetweenLidarScans = yaw - self.previousYaw
+        if self.readFrom == READ_FROM_SERIAL:
+            yaw = None        
+            while self.pipeFromImu.poll():
+                yaw = self.pipeFromImu.recv()
+            if yaw is not None:
+                if self.previousYaw is not None:
+                    imuYawBetweenLidarScans = yaw - self.previousYaw
+                    self.previousYaw = yaw
+                    return imuYawBetweenLidarScans
                 self.previousYaw = yaw
-                return imuYawBetweenLidarScans
-            self.previousYaw = yaw
-        return None
+            return None
+        else:
+            return 0
+
+    def getYawSum(self):
+        return self.previousYaw
 
 if __name__ == '__main__':
     px = pixhawk()
